+++ replit.nix
@@ -1,5 +1,9 @@
 { pkgs }: {
 	deps = [
+   pkgs.fuse-emulator
+   pkgs.kmod
+   pkgs.sshfs
+   pkgs.openssh
 		pkgs.clang
 		pkgs.ccls
 		pkgs.gdb