+++ 👨‍💻/lib/GRP.hpp
@@ -5,31 +5,25 @@
 #include <sstream>
 
 namespace GRP {
-	using integer = mpz_class;
-	using flat = mpf_class;
-	using number = mpq_class;
-
-	using str = std::string;
-
-	number pow(number base, integer exp) {
-		number result = 1;
-		for (integer i = 0; i < exp; i++)
+	mpq_class pow(mpq_class base, mpz_class exp) {
+		mpq_class result = 1;
+		for (mpz_class i = 0; i < exp; i++)
 			result *= base;
 		return result;
 	}
 
-	integer precision;
-	integer kilo = 1'000;
+	mpz_class precision;
+	mpz_class kilo = 1'000;
 
-	std::vector<str> formatLong = {" thousand"," million"," billion"," trillion"," quadrillion"," quintillion"," sextillion"," septillion"," octillion"," nonillion"};
-	std::vector<str> prefixes = {"","un","duo","tre","quattuor","quin","sex","septen","octo","novem"};
-	std::vector<str> suffixes = {"decillion","vigintillion","trigintillion","quadragintillion","quinquagintillion","sexagintillion","septuagintillion","octogintillion","nonagintillion"};
+	std::vector<std::string> formatLong = {" thousand"," million"," billion"," trillion"," quadrillion"," quintillion"," sextillion"," septillion"," octillion"," nonillion"};
+	std::vector<std::string> prefixes = {"","un","duo","tre","quattuor","quin","sex","septen","octo","novem"};
+	std::vector<std::string> suffixes = {"decillion","vigintillion","trigintillion","quadragintillion","quinquagintillion","sexagintillion","septuagintillion","octogintillion","nonagintillion"};
 
-	integer round(number x) {
+	mpz_class round(mpq_class x) {
 		return  x.get_num() / x.get_den();
 	}
 
-	str toString(number x) {
+	std::string toString(mpq_class x) {
 		std::stringstream result;
 		if (x >= 1'000'000) {
 			size_t suffix = 1;
@@ -43,13 +37,13 @@ namespace GRP {
 				suffix++;
 			}
 
-			integer big = round(x / kilo);
-			integer small = round(x - big * kilo);
+			mpz_class big = round(x / kilo);
+			mpz_class small = round(x - big * kilo);
 
 			if(small == 0) {
 				result << big << formatLong[suffix];
 			} else {
-				str sml = small.get_str();
+				std::string sml = small.get_str();
 				char smallf[] = "000";
 				for(int i = 0; i < sml.length(); i++) {
 					smallf[2 - i] = sml[sml.length() - i - 1];
@@ -64,14 +58,14 @@ namespace GRP {
 			return result.str();
 		} else {
 			x *= 10;
-			integer big = round(x/10);
+			mpz_class big = round(x/10);
 
-			integer small = round(x - big * 10);
+			mpz_class small = round(x - big * 10);
 
-			integer kilos = big / kilo;
-			integer ones = big - kilos * kilo;
+			mpz_class kilos = big / kilo;
+			mpz_class ones = big - kilos * kilo;
 			if(kilos > 0) {
-				str oness = ones.get_str();
+				std::string oness = ones.get_str();
 				char onesf[] = "000";
 				for(size_t i = 0; i < oness.length(); i++) {
 					onesf[2 - i] = oness[oness.length() - i - 1];
@@ -89,7 +83,7 @@ namespace GRP {
 		return "error: HOW!";
 	}
 
-	void init(integer precision = 1) {
+	void init(mpz_class precision = 1) {
 		precision = pow(precision, 10);
 
 		for(int i = 0; i < suffixes.size(); i++) {