+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [Unreleased]
+
+### added
+
+- triggers
+- building unlocking
+
 ## [0.0.1_2] 2024-08-07
 
 ### fixed